drivers/spi: Resolve a Compile Warning in the Axxia Driver
Upgrading to GCC 5.3.0 produced a warning in the Axxia SPI
driver.  This commit fixes the problem.

+++ drivers/spi/axxia_spi.c
@@ -387,7 +387,7 @@ spi_xfer(struct spi_slave *slave, unsigned int bitlen,
 	const u8 *txp = dout;
 	u8 *rxp = din;
 #ifdef ENABLE_DMA
-	unsigned long address = NULL;
+	unsigned long address = (unsigned long)NULL;
 	unsigned int direction = 0;
 #endif	/* ENABLE_DMA */
